Add SQL for cleanup into regression test.
Previously when regression test was repeated, the second and
subsequent tests always failed because the roles created by
the first test could cause them to emit an error.

To be able to repeat the test successfully, this commit adds
the SQL commands that drop the roles and all the objects
owned by them, into the end of regression test.