Temporarily disable postgres_fdw_plus regression test for parallel_commit=on.
Commit f702537 introduced a regression test to verify that postgres_fdw_plus
works with parallel_commit enabled. However, this test has unexpectedly
caused instability in the test suite.

To allow continued development of postgres_fdw_plus, this commit temporarily
disables the regression test for parallel_commit=on. The regression test for
parallel_commit=on will be re-enabled once it has been improved and stabilized.