Refactor code for prompt return in pgfdw_xact_two_phase().
This commit simplifies the code in pgfdw_xact_two_phase() by refactoring
the handling of cases where two-phase commit is not enabled or when
the PREPARE command is executed.  In such scenarios, pgfdw_xact_two_phase()
should return promptly, and the responsibility of handling these cases is
delegated to its caller, pgfdw_xact_callback(). The refactoring in
this commit improves the readability and maintainability of the code in
postgres_fdw_plus related to these scenarios.

Author: @MasaoFujii 
Reviewed-by: @YKatsuragi 
Discussion: https://github.com/pgfdwplus/postgres_fdw_plus/pull/12